separated styles, added comments, animate controls
