+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { Home, VideoScreen } from './src/components/screens'
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
+import { ScreenNames } from './src/utils'
 const Stack = createNativeStackNavigator();
 
 function App(props) {
@@ -11,8 +11,8 @@ function App(props) {
         <Stack.Navigator screenOptions={{
           headerShown: false
         }}>
-        <Stack.Screen name="none" component={Home} />
-        <Stack.Screen name="videoScreen" component={VideoScreen} />
+        <Stack.Screen name="HOME" component={Home} />
+        <Stack.Screen name={ScreenNames.VIDEOSCREEN} component={VideoScreen} />
       </Stack.Navigator>
       </NavigationContainer>
   );
