+++ src/components/commoncomponents/index.js
@@ -1,5 +1,4 @@
 
 import VideoPlayer from './videoPlayer'
-import ModalContainer from './modalContainer'
 import Header from './header'
-export { VideoPlayer, ModalContainer, Header };
+export { VideoPlayer, Header };
