--- src/components/commoncomponents/modalContainer/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { View } from 'react-native'
-
-function ModelContainer(props) {
-    return (
-       <View>
-
-       </View>
-    );
-}
-
-export default ModelContainer;
