+++ src/components/commoncomponents/videoPlayer/index.js
@@ -1,33 +1,45 @@
 import React, { useRef, useState } from 'react';
-import { Text, View, StyleSheet, Image, TouchableOpacity } from 'react-native'
+import { View, Image, TouchableOpacity, TouchableHighlight } from 'react-native'
 import Video from 'react-native-video';
-import {widthPercentageToDP as wp, heightPercentageToDP as hp} from 'react-native-responsive-screen';
-
-
-
+import * as Animatable from 'react-native-animatable'
+import { Header } from './..'
+import styles from './styles'
+/*
+    Video Player Component to play video based on URL which will come as a prop.
+*/
 function VideoPlayer(props) {
 
-    const [ totalDuration ,setTotalDuration ] = useState(0)
-    const [ currentTime ,setCurrentTime ] = useState(0)
+    const [ totalDuration ,setTotalDuration ] = useState(0) // Total video duration
+    const [ currentTime ,setCurrentTime ] = useState(0)  // current video playing time
+    const [ openControls ,setopenControls ] = useState(false) // Open control whever click on video.
+
     let player = useRef(null)
 
+    //On video load. this method will call to capture the video length.
     function onLoad(data){
         setTotalDuration(data.duration)
     }
+
+    //While video running this function will call to get current video time.
     function onProgress(data){
         setCurrentTime(data.currentTime)
     }
+
+    //This function will check current video time is 10 sec greater than total video length. If true than it will backward the video.
     function forward(){
-        if(currentTime+10<=totalDuration) player.seek(currentTime+10) 
+        if(currentTime+10<=totalDuration) player.seek(currentTime+10)
     }
+
+    //This function will check current video time is 10 sec less than total video length. If true than it will backward the video.
     function backword(){
         if(currentTime-10<=totalDuration) player.seek(currentTime-10) 
     }
 
-    let { openControls, isPause, URL, videoPaused } = props;
+    let { isPause, URL, videoPaused, navigation } = props;
     return (
-        <View>
-                <>
+        <View style={styles.container} >
+            <TouchableHighlight onPress={() =>  setopenControls(!openControls) }>
+            <>
                 <Video source={URL}
                     ref={(ref) => {
                         player = ref
@@ -37,9 +49,13 @@ function VideoPlayer(props) {
                     fullscreen={true}              
                     resizeMode="stretch"
                     onProgress={onProgress.bind(this)}
-                    style={styles.backgroundVideo} 
+                    style={styles.backgroundVideo}
                     />
-                    {openControls ? <TouchableOpacity activeOpacity={1} style={styles.controls}>
+                   <>
+                        <Animatable.View animation={openControls ? "slideOutDown" : "slideOutUp"} style={styles.topControls}>
+                            <Header onClick ={()=> navigation.goBack(null)} />
+                        </Animatable.View>
+                        <Animatable.View animation={!openControls ? "slideOutDown" : "slideOutUp"} style={styles.controls}>
                         <TouchableOpacity onPress={()=> backword()}>
                             <Image style={{...styles.image, transform:[{ rotate: '180deg'}]}} source={require("./../../../assets/forward.png")} />
                         </TouchableOpacity>
@@ -52,62 +68,13 @@ function VideoPlayer(props) {
                         <TouchableOpacity onPress={()=> forward()}>
                             <Image style={styles.image} source={require("./../../../assets/forward.png")} />
                         </TouchableOpacity>
-                    </TouchableOpacity> : null}
+                    </Animatable.View> 
+                    </>
                 </>
-            </View>
+            </TouchableHighlight>
+             </View>
     );
 }
-
-
-
-const styles = StyleSheet.create({
-    container:{
-        flex: 1
-    },
-    backgroundVideo: {
-        width: '100%',
-        height: '100%'
-    },
-    sliderControl:{
-        flex:1,
-        flexDirection:'row'
-    },
-    controls:{
-        position:'absolute',
-        flexDirection:'row',
-        alignItems:'center',
-        justifyContent:'space-evenly',
-        width: '100%',
-        height: wp("15"),
-        bottom: 0,
-        backgroundColor: 'rgba(0,0,0,0.5)'
-    },
-    image:{
-        width: wp("8"),
-        height: wp("8")
-    },
-    animateButton:{
-        position:'absolute',
-        flexDirection:'row',
-        width:'100%',
-        justifyContent:'center',
-        bottom: hp("20")
-    },
-    resIcon:{
-        width: wp("6"),
-        height: wp("6")
-    },
-    animatedView:{
-        width: wp("12"),
-        height:wp("12"),
-        marginRight: wp("2"),
-        borderRadius: wp("12")/2,
-        justifyContent:'center',
-        alignItems:'center',
-        backgroundColor: 'white',
-        elevation:5
-    }
-})
 
 
 export default VideoPlayer;
