+++ src/utils/index.js
@@ -1,0 +1,4 @@
+import logger from './logger'
+import ScreenNames from './screenNames'
+
+export { logger, ScreenNames }
