+++ src/utils/logger.js
@@ -1,0 +1,6 @@
+
+//Logger Functionality for dev
+export default logger ={
+    log: __DEV__ ? console.log : () => {},
+    error: __DEV__ ? console.log : () => {},
+};
