+++ src/utils/screenNames.js
@@ -1,0 +1,3 @@
+export default ScreenNames = {
+    VIDEOSCREEN: "VIDEOSCREEN"
+}
